+++ packages/credential-sdk/src/types/did/document/const.js
@@ -1,0 +1,3 @@
+export const ATTESTS_IRI = 'https://rdf.dock.io/alpha/2021#attestsDocumentContents';
+
+export const CONTEXT_URI = 'https://www.w3.org/ns/did/v1';