+++ packages/credential-sdk/src/types/generic/with-base58.js
@@ -1,0 +1,19 @@
+import withFrom from './with-from';
+import { decodeFromBase58, encodeAsBase58 } from '../../utils/base-x';
+
+export default function withBase58(klass) {
+  const name = `withBase58(${klass.name})`;
+  const obj = {
+    [name]: class extends klass {
+      get value() {
+        return String(this);
+      }
+
+      toString() {
+        return encodeAsBase58(this.bytes);
+      }
+    },
+  };
+
+  return withFrom(obj[name], (value, from) => from(typeof value === 'string' ? decodeFromBase58(value) : value));
+}