+++ packages/credential-sdk/src/types/generic/with-base64.js
@@ -1,0 +1,15 @@
+import withFrom from './with-from';
+import { decodeFromBase64, encodeAsBase64 } from '../../utils/base-x';
+
+export default function withBase64(klass) {
+  const name = `withBase64(${klass.name})`;
+  const obj = {
+    [name]: class extends klass {
+      toString() {
+        return encodeAsBase64(this.bytes);
+      }
+    },
+  };
+
+  return withFrom(obj[name], (value, from) => from(typeof value === 'string' ? decodeFromBase64(value) : value));
+}