+++ packages/cheqd-blockchain-modules/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @docknetwork/cheqd-blockchain-modules
 
+## 0.10.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.13.0
+
 ## 0.10.0
 
 ### Minor Changes