+++ packages/dock-blockchain-api/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @docknetwork/dock-blockchain-api
 
+## 0.7.0
+
+### Minor Changes
+
+- Small refactoring of the `did/document`
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.13.0
+
 ## 0.6.1
 
 ### Patch Changes