+++ packages/dock-blockchain-modules/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @docknetwork/dock-blockchain-modules
 
+## 0.8.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.13.0
+
 ## 0.8.0
 
 ### Minor Changes